pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "compound-agent"
version = "0.1.0"
description = "Autonomous knowledge management agent with multi-agent orchestration"
requires-python = ">=3.11"
dependencies = [
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.100.0",
    "google-genai>=1.0.0",
    "beautifulsoup4>=4.12.0",
    "apscheduler>=3.10.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "defusedxml>=0.7.0",
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
